backend/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.25.0

# Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.6

# Typing
typing-extensions>=4.8.0
